refactor(knowledge): move findSources/compareClaims/nextMove rankings into Knowledge (#366)
Relocate three pure, Obsidian-free ranking projections out of `actions/` into the
Knowledge Model, hardening the Commands→Queries boundary (constitution §XI):

- findSourcesLogic, compareClaimsLogic → architecture/knowledge/query
- nextMoveLogic → architecture/knowledge/state

Owning actions now import them from Knowledge (the allowed engine→knowledge
direction); no shim aliases are left behind. The registration guards and the
pure-respects-direction boundary test are repointed to the new homes, with a
regression pin that keeps `query` + `state` in the scanned set. Zero behaviour
change — the characterisation suites pass identically. verify + lint:obsidian green.
