+++ docs/architecture/reposition-map.md
@@ -25,10 +25,12 @@ never `application`/`zettelkasten`/`hooks`/`config`/`starters`/`actions`/`archit
 `architecture/plugin`/`architecture/api`/`obsidian` (the one allowed exception is the zero-import type
 `architecture/plugin/model/FrontmatterModel`). This is enforced by
 `test/architecture/knowledge/pure-respects-direction.test.ts` (alongside the obsidian-free grep). To
-make it hold, four pure "query" logics that had drifted into `actions/` were repositioned back into
-Knowledge — `maturityLogic → derive`, `relationRankingLogic → relations`, and
-`findContradictionLogic`/`findUnansweredQuestionLogic → query` — so their owning actions now import
-them from Knowledge (the correct engine→knowledge direction), with zero behaviour change. The
+make it hold, seven pure "query" logics that had drifted into `actions/` were repositioned back into
+Knowledge — `maturityLogic → derive`, `relationRankingLogic → relations`,
+`findContradictionLogic`/`findUnansweredQuestionLogic → query` (#209), and then
+`findSourcesLogic`/`compareClaimsLogic → query` plus `nextMoveLogic → state` (#366, D6) — so their
+owning actions now import them from Knowledge (the correct engine→knowledge direction), with zero
+behaviour change. The
 `KnowledgeIndex`/`snapshot` root adapters legitimately touch Obsidian and stay outside the pure gate.
 
 ## Workflow Engine — turning a canvas into a note
