+++ src/architecture/knowledge/query/compareClaimsLogic.ts
