+++ src/architecture/knowledge/state/nextMoveLogic.ts
